Global build attributes
